% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBP_aggregate_fischer.R
\name{NBP_aggregate_fischer}
\alias{NBP_aggregate_fischer}
\title{Aggregation of p values with Fischer's method}
\usage{
NBP_aggregate_fischer(p_vals, evidence.limit = 6.33e-05)
}
\arguments{
\item{p_vals}{A vector of p-values}

\item{evidence.limit}{A threshold for the minimum p-value. This is used
before aggregation to ensure that no outlier p-value completely dominates
the overall aggregated p-value.
Up to 4 sd out for Gaussian: 1-2\emph{abs(pnorm(4)-0.5) = 6.33e-05.
Two of these aggregated with the function: 8.1E-8.
Up to 3.5 sd out for Gaussian: 1-2}abs(pnorm(3.5)-0.5) = 4.7E-4.
Two of these aggregated with the function: 3.6E-6.}
}
\value{
One p-value
}
\description{
Aggregation of p values with Fischer's method
}
\examples{
NBP_aggregate_fischer(c(1E-3))
NBP_aggregate_fischer(c(1E-3, 1E-3))
NBP_aggregate_fischer(c(1E-3, 1E-3, 1E-3))
NBP_aggregate_fischer(c(1E-3, 1E-3, 1E-3, 1, 1))


}
