% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBP_aggregate_p_values.R
\name{NBP_aggregate_p_values}
\alias{NBP_aggregate_p_values}
\title{Aggregate for results of NBP_classify}
\usage{
NBP_aggregate_p_values(p_val, by, type = "fischer", evidence.limit = 6.33e-05)
}
\arguments{
\item{p_val}{A vector of p-values}

\item{by}{A dataframe of columns to aggregate by, similar to the use in the
aggregate function.}

\item{type}{Type of aggregation.}

\item{evidence.limit}{See documentation of NBP_aggregate_fischer}
}
\value{
A single p-value
}
\description{
This function shows the use of aggregate for dataframes that result from
the function NBP_classify. You can either use this wrapper, or use aggregate
or dplyr or other.
}
\examples{
NBP_aggregate_p_values(c(1E-3, 1E-3, 1E-3, 1E-3, 1), by=data.frame(by1=c(1,2,3,2,3)))


}
