% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBP_classify.R
\name{NBP_classify}
\alias{NBP_classify}
\title{Classification with NBP}
\usage{
NBP_classify(
  train_cov,
  train_class,
  predict_cov,
  keep = FALSE,
  family = "Gaussian",
  identifiers = NULL
)
}
\arguments{
\item{train_cov}{Covariates for the training data}

\item{train_class}{Class membership (response category) for the training data.
Must be the same length as number of rows in train_cov.}

\item{predict_cov}{Covariates for the prediction, where the class is not known.}

\item{keep}{Do you want the density functions for plotting?}

\item{family}{The distribution family for the covariates.
Only Gaussian is implemented.}

\item{identifiers}{Set of identifiers or metadata, belonging to the prediction covariates.
Will be added to the output dataframe.}
}
\value{
A dataframe with p-values for predicting using the predict_cov at
each of the different possible classes.
}
\description{
NBP is Naive Bayes with p-values/exceedance probabilities, unlike the standard Naive
Bayes which uses relative probability densities. However, this algorithm
can also be used to perform Naive Bayes.
See also Arxiv paper (TODO).
}
\examples{
NBP_classify(iris[-c(1,2), 1:4], iris[-c(1,2), 5], iris[c(1,2), 1:4])


}
