% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBP_pred_categories.R
\name{NBP_pred_categories}
\alias{NBP_pred_categories}
\title{Turn p-values into interpretable categories}
\usage{
NBP_pred_categories(
  dat,
  type = "Avoid_0pred_V1",
  cutoff = 0,
  do_checks = FALSE
)
}
\arguments{
\item{dat}{A dataframe with all numeric columns being p-values}

\item{type}{The type of categorisation you want. See code comments
inside the
function for details.}

\item{cutoff}{Some types require you to set a cutoff}

\item{do_checks}{Should extra checks be done if possible? Can result in
warnings that are helpful/unhelpful.}
}
\value{
The original dataframe but where all numerical values have been
replaced by the thresholded values.
}
\description{
This function implements simple logic on how to threshold continuous values
into discrete values.
}
\examples{
NBP_pred_categories(c(0.1, 1E-10, 0, 1))

}
